@@ -1,9 +1,8 @@
 package context
 
 import (
-	"encoding/base32"
 	"encoding/base64"
-	"encoding/json"
+	"fmt"
 	"github.com/google/uuid"
 	"net/http"
 )
@@ -12,41 +11,50 @@ type Id uuid.UUID
 
 var (
 	NilRequestId = Id(uuid.Nil)
-	idEncoding   = base64.StdEncoding.WithPadding(base32.NoPadding)
+	idEncoding   = base64.StdEncoding.WithPadding(base64.NoPadding)
 )
 
-func NewId(fromOtherReverseProxy bool, req *http.Request) Id {
+func NewId(fromOtherReverseProxy bool, req *http.Request) (Id, error) {
 	return newId(fromOtherReverseProxy, "X-Request-ID", req)
 }
 
-func NewCorrelationId(fromOtherReverseProxy bool, req *http.Request) Id {
-	return newId(fromOtherReverseProxy, "X-Correlation-ID", req)
+func NewCorrelationId(req *http.Request) (Id, error) {
+	return newId(true, "X-Correlation-ID", req)
 }
 
-func newId(fromOtherReverseProxy bool, fromHeader string, req *http.Request) Id {
-	if fromOtherReverseProxy {
+func newId(acceptUpstreamHeaderIfAny bool, fromHeader string, req *http.Request) (Id, error) {
+	if acceptUpstreamHeaderIfAny {
 		if x := req.Header.Get(fromHeader); len(x) > 0 && len(x) <= 256 {
-			if id, err := ParseId(x); err == nil {
-				return id
+			var id Id
+			if err := id.UnmarshalText([]byte(x)); err == nil {
+				return id, nil
 			}
 		}
 	}
-	val := uuid.New()
-	return Id(val)
-}
-
-func (instance Id) String() string {
-	return idEncoding.EncodeToString(instance[:])
-}
-
-func (instance Id) MarshalJSON() ([]byte, error) {
-	return json.Marshal(instance.String())
-}
-
-func ParseId(plain string) (Id, error) {
-	val, err := uuid.Parse(plain)
+	val, err := uuid.NewRandom()
 	if err != nil {
 		return Id{}, err
 	}
 	return Id(val), nil
+}
+
+func (this Id) String() string {
+	return idEncoding.EncodeToString(this[:])
+}
+
+func (this Id) MarshalText() ([]byte, error) {
+	return []byte(this.String()), nil
+}
+
+func (this *Id) UnmarshalText(in []byte) error {
+	decoded, err := idEncoding.DecodeString(string(in))
+	if err != nil {
+		return fmt.Errorf("illegal id: %q", string(in))
+	}
+	var buf uuid.UUID
+	if err := buf.UnmarshalBinary(decoded); err != nil {
+		return fmt.Errorf("illegal id: %q", string(in))
+	}
+	*this = Id(buf)
+	return nil
 }
