@@ -1,7 +1,6 @@
 package context
 
 import (
-	"errors"
 	"fmt"
 )
 
@@ -19,8 +18,6 @@ const (
 )
 
 var (
-	ErrIllegalStage = errors.New("illegal stage")
-
 	stageToName = map[Stage]string{
 		StageUnknown:                "unknown",
 		StageCreated:                "created",
@@ -31,28 +28,12 @@ var (
 		StageSendResponseToClient:   "sendResponseToClient",
 		StageDone:                   "done",
 	}
-
-	nameToStage = func(in map[Stage]string) map[string]Stage {
-		out := make(map[string]Stage)
-		for stage, name := range in {
-			out[name] = stage
-		}
-		return out
-	}(stageToName)
 )
 
-func ParseStage(plain string) (Stage, error) {
-	if stage, ok := nameToStage[plain]; ok {
-		return stage, nil
-	} else {
-		return 0, ErrIllegalStage
-	}
-}
-
-func (instance Stage) String() string {
-	if name, ok := stageToName[instance]; ok {
+func (this Stage) String() string {
+	if name, ok := stageToName[this]; ok {
 		return name
 	} else {
-		return fmt.Sprintf("unknown-stage-%d", instance)
+		return fmt.Sprintf("unknown-stage-%d", this)
 	}
 }
