@@ -2,7 +2,9 @@ package definition
 
 import (
 	"fmt"
+	"github.com/echocat/lingress/settings"
 	"github.com/echocat/lingress/support"
+	log "github.com/echocat/slf4g"
 	"k8s.io/client-go/kubernetes"
 	"time"
 )
@@ -13,12 +15,12 @@ type Definitions struct {
 	Service        *Service
 }
 
-func New(client kubernetes.Interface, resyncAfter time.Duration) (*Definitions, error) {
-	if serviceSecrets, err := NewServiceSecrets(client, resyncAfter); err != nil {
+func New(s *settings.Settings, client kubernetes.Interface, resyncAfter time.Duration, logger log.Logger) (*Definitions, error) {
+	if serviceSecrets, err := NewServiceSecrets(s, client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create service secrets definition store: %v", err)
-	} else if ingress, err := NewIngress(client, resyncAfter); err != nil {
+	} else if ingress, err := NewIngress(client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create ingress definition store: %v", err)
-	} else if service, err := NewService(client, resyncAfter); err != nil {
+	} else if service, err := NewService(client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create service definition store: %v", err)
 	} else {
 		return &Definitions{
@@ -29,28 +31,28 @@ func New(client kubernetes.Interface, resyncAfter time.Duration) (*Definitions, 
 	}
 }
 
-func (instance *Definitions) SetNamespace(namespace string) {
-	instance.ServiceSecrets.SetNamespace(namespace)
+func (this *Definitions) SetNamespace(namespace string) {
+	this.ServiceSecrets.SetNamespace(namespace)
 }
 
-func (instance *Definitions) Init(stop support.Channel) error {
-	if err := instance.ServiceSecrets.Init(stop); err != nil {
+func (this *Definitions) Init(stop support.Channel) error {
+	if err := this.ServiceSecrets.Init(stop); err != nil {
 		return err
 	}
 
-	if err := instance.Service.Init(stop); err != nil {
+	if err := this.Service.Init(stop); err != nil {
 		return err
 	}
 
-	if err := instance.Ingress.Init(stop); err != nil {
+	if err := this.Ingress.Init(stop); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func (instance *Definitions) HasSynced() bool {
-	return instance.Ingress.HasSynced() &&
-		instance.Service.HasSynced() &&
-		instance.ServiceSecrets.HasSynced()
+func (this *Definitions) HasSynced() bool {
+	return this.Ingress.HasSynced() &&
+		this.Service.HasSynced() &&
+		this.ServiceSecrets.HasSynced()
 }
