@@ -5,10 +5,12 @@ import (
 	"fmt"
 	lctx "github.com/echocat/lingress/context"
 	"github.com/echocat/lingress/fallback"
+	"github.com/echocat/lingress/file/providers"
 	"github.com/echocat/lingress/management"
 	"github.com/echocat/lingress/proxy"
 	"github.com/echocat/lingress/rules"
 	"github.com/echocat/lingress/server"
+	"github.com/echocat/lingress/settings"
 	"github.com/echocat/lingress/support"
 	"github.com/echocat/slf4g"
 	"github.com/echocat/slf4g/level"
@@ -17,22 +19,25 @@ import (
 	"reflect"
 	"strings"
 	"sync/atomic"
-	"time"
 )
 
 type Lingress struct {
-	RulesRepository    rules.CombinedRepository
-	Proxy              *proxy.Proxy
-	Fallback           *fallback.Fallback
-	Management         *management.Management
-	Http               *server.HttpConnector
-	Https              *server.HttpConnector
-	AccessLogQueueSize uint16
-	InlineAccessLog    bool
+	settings *settings.Settings
+
+	RulesRepository rules.CombinedRepository
+	Proxy           *proxy.Proxy
+	Fallback        *fallback.Fallback
+	Management      *management.Management
+	Http            *server.HttpConnector
+	Https           *server.HttpConnector
 
 	accessLogQueue chan *accessLogEntry
 
 	unprocessableConnectionDocumented map[reflect.Type]bool
+
+	logger                 log.Logger
+	accessLogger           log.Logger
+	accessLoggerMessageKey string
 }
 
 type ConnectionStates struct {
@@ -41,89 +46,76 @@ type ConnectionStates struct {
 	Idle   uint64
 }
 
-func New(fps support.FileProviders) (*Lingress, error) {
+func New(s *settings.Settings, fps providers.FileProviders) (*Lingress, error) {
+	logProvider := log.GetProvider()
+
 	connectorIds := []server.ConnectorId{server.DefaultConnectorIdHttp, server.DefaultConnectorIdHttps}
-	if r, err := rules.NewRepository(); err != nil {
+	r, err := rules.NewRepository(s, logProvider.GetLogger("rules"))
+	if err != nil {
 		return nil, err
-	} else if p, err := proxy.New(r); err != nil {
-		return nil, err
-	} else if f, err := fallback.New(fps); err != nil {
-		return nil, err
-	} else if m, err := management.New(connectorIds, r); err != nil {
-		return nil, err
-	} else if hHttp, err := server.NewHttpConnector(server.DefaultConnectorIdHttp); err != nil {
-		return nil, err
-	} else if hHttps, err := server.NewHttpConnector(server.DefaultConnectorIdHttps); err != nil {
-		return nil, err
-	} else {
-		result := &Lingress{
-			RulesRepository:    r,
-			Proxy:              p,
-			Fallback:           f,
-			Management:         m,
-			Http:               hHttp,
-			Https:              hHttps,
-			AccessLogQueueSize: 5000,
-		}
-
-		result.Http.Handler = result
-		result.Http.MaxConnections = 256
-
-		result.Https.Handler = result
-		result.Https.Server.Addr = ":8443"
-
-		p.ResultHandler = result.onResult
-		p.AccessLogger = result.onAccessLog
-		p.MetricsCollector = result.Management
-
-		return result, nil
 	}
+	p, err := proxy.New(s, r, logProvider.GetLogger("proxy"))
+	if err != nil {
+		return nil, err
+	}
+	f, err := fallback.New(s, fps, logProvider.GetLogger("fallback"))
+	if err != nil {
+		return nil, err
+	}
+	m, err := management.New(s, connectorIds, r, logProvider.GetLogger("management"))
+	if err != nil {
+		return nil, err
+	}
+	hHttp, err := server.NewHttpConnector(s, server.DefaultConnectorIdHttp, logProvider.GetLogger(string(server.DefaultConnectorIdHttp)))
+	if err != nil {
+		return nil, err
+	}
+	hHttps, err := server.NewHttpConnector(s, server.DefaultConnectorIdHttps, logProvider.GetLogger(string(server.DefaultConnectorIdHttps)))
+	if err != nil {
+		return nil, err
+	}
+
+	result := &Lingress{
+		settings: s,
+
+		RulesRepository: r,
+		Proxy:           p,
+		Fallback:        f,
+		Management:      m,
+		Http:            hHttp,
+		Https:           hHttps,
+
+		accessLogger: logProvider.GetLogger("accessLog"),
+		logger:       logProvider.GetLogger("core"),
+	}
+
+	result.accessLoggerMessageKey = result.accessLogger.GetProvider().GetFieldKeysSpec().GetMessage()
+	result.Http.Handler = result
+	result.Https.Handler = result
+
+	p.ResultHandler = result.onResult
+	p.AccessLogger = result.onAccessLog
+	p.MetricsCollector = result.Management
+
+	return result, nil
 }
 
-func (instance *Lingress) RegisterFlag(fe support.FlagEnabled, appPrefix string) error {
-	fe.Flag("accessLog.queueSize", "Maximum number of accessLog elements that could be queue before blocking.").
-		PlaceHolder(fmt.Sprint(instance.AccessLogQueueSize)).
-		Envar(support.FlagEnvName(appPrefix, "ACCESS_LOG_QUEUE_SIZE")).
-		Uint16Var(&instance.AccessLogQueueSize)
-	fe.Flag("accessLog.inline", "Instead of exploding the accessLog entries into sub-entries everything is inlined into the root object.").
-		Envar(support.FlagEnvName(appPrefix, "ACCESS_LOG_INLINE")).
-		BoolVar(&instance.InlineAccessLog)
-	if err := instance.RulesRepository.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	if err := instance.Proxy.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	if err := instance.Fallback.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	if err := instance.Management.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	if err := instance.Http.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	if err := instance.Https.RegisterFlag(fe, appPrefix); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (instance *Lingress) ServeHTTP(connector server.Connector, resp http.ResponseWriter, req *http.Request) {
-	finalize := instance.Management.CollectClientStarted(connector.GetId())
+func (this *Lingress) ServeHTTP(connector server.Connector, resp http.ResponseWriter, req *http.Request) {
+	finalize := this.Management.CollectClientStarted(connector.GetId())
 	defer finalize()
 
-	instance.Proxy.ServeHTTP(connector, resp, req)
+	this.Proxy.ServeHTTP(connector, resp, req)
 }
 
-func (instance *Lingress) OnConnState(connector server.Connector, conn net.Conn, state http.ConnState) {
+func (this *Lingress) OnConnState(connector server.Connector, conn net.Conn, state http.ConnState) {
 	annotated, ok := conn.RemoteAddr().(server.AnnotatedAddr)
 	if !ok {
 		connType := reflect.TypeOf(conn)
-		if !instance.unprocessableConnectionDocumented[connType] {
-			instance.unprocessableConnectionDocumented[connType] = true
-			log.With("connType", connType.String()).
-				Error("cannot inspect connection of provided connection type; for those kinds of connections there will be no statistics be available")
+		if !this.unprocessableConnectionDocumented[connType] {
+			this.unprocessableConnectionDocumented[connType] = true
+			this.logger.
+				With("connType", connType.String()).
+				Error("Cannot inspect connection of provided connection type; for those kinds of connections there will be no statistics be available.")
 		}
 		return
 	}
@@ -135,7 +127,7 @@ func (instance *Lingress) OnConnState(connector server.Connector, conn net.Conn,
 		return
 	}
 
-	client, ok := instance.Management.Metrics.Client[connector.GetId()]
+	client, ok := this.Management.Metrics.Client[connector.GetId()]
 	if !ok {
 		return
 	}
@@ -170,67 +162,74 @@ func (instance *Lingress) OnConnState(connector server.Connector, conn net.Conn,
 	return
 }
 
-func (instance *Lingress) Init(stop support.Channel) error {
-	if err := instance.RulesRepository.Init(stop); err != nil {
+func (this *Lingress) Init(stop support.Channel) error {
+	if err := this.RulesRepository.Init(stop); err != nil {
 		return err
 	}
-	if err := instance.Proxy.Init(stop); err != nil {
+	if err := this.Proxy.Init(stop); err != nil {
 		return err
 	}
 
-	if s := instance.AccessLogQueueSize; s > 0 {
-		queue := make(chan *accessLogEntry, instance.AccessLogQueueSize)
-		go instance.accessLogQueueWorker(stop, queue)
-		instance.accessLogQueue = queue
+	if s := this.settings.AccessLog.QueueSize; s > 0 {
+		queue := make(chan *accessLogEntry, this.settings.AccessLog.QueueSize)
+		go this.accessLogQueueWorker(stop, queue)
+		this.accessLogQueue = queue
 	} else {
-		instance.accessLogQueue = nil
+		this.accessLogQueue = nil
 	}
 
-	instance.unprocessableConnectionDocumented = make(map[reflect.Type]bool)
+	this.unprocessableConnectionDocumented = make(map[reflect.Type]bool)
 
-	go instance.shutdownListener(stop)
+	go this.shutdownListener(stop)
 
-	if tlsConfig, err := instance.createTlsConfig(); err != nil {
+	if tlsConfig, err := this.createTlsConfig(); err != nil {
 		return err
 	} else {
-		instance.Https.Server.TLSConfig = tlsConfig
+		this.Https.Server.TLSConfig = tlsConfig
 	}
 
-	if err := instance.Http.Serve(stop); err != nil {
+	if err := this.Http.Serve(stop); err != nil {
 		return err
 	}
-	if err := instance.Https.Serve(stop); err != nil {
+	if err := this.Https.Serve(stop); err != nil {
 		return err
 	}
 
-	if err := instance.Management.Init(stop); err != nil {
+	if err := this.Management.Init(stop); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func (instance *Lingress) createTlsConfig() (*tls.Config, error) {
+func (this *Lingress) createTlsConfig() (*tls.Config, error) {
 	fail := func(err error) (*tls.Config, error) {
 		return nil, fmt.Errorf("cannot create TLS config: %w", err)
 	}
-	defaultCert, err := support.CreateDummyCertificate()
-	if err != nil {
-		return fail(err)
+
+	result := tls.Config{
+		Certificates:   []tls.Certificate{},
+		GetCertificate: this.resolveCertificate,
 	}
 
-	return &tls.Config{
-		Certificates: []tls.Certificate{
-			defaultCert,
-		},
-		GetCertificate: instance.resolveCertificate,
-	}, nil
+	if this.settings.Tls.FallbackCertificate.Get() {
+		v, err := support.CreateDummyCertificate()
+		if err != nil {
+			return fail(err)
+		}
+		result.Certificates = append(result.Certificates, v)
+	}
+
+	return &result, nil
 }
 
-func (instance *Lingress) resolveCertificate(info *tls.ClientHelloInfo) (*tls.Certificate, error) {
-	certificates, err := instance.RulesRepository.FindCertificatesBy(rules.CertificateQuery{
-		Host: info.ServerName,
-	})
+func (this *Lingress) resolveCertificate(info *tls.ClientHelloInfo) (*tls.Certificate, error) {
+	var query rules.CertificateQuery
+	if err := query.Host.Set(info.ServerName); err != nil {
+		return nil, nil
+	}
+
+	certificates, err := this.RulesRepository.FindCertificatesBy(query)
 	if err != nil {
 		return nil, err
 	}
@@ -244,41 +243,47 @@ func (instance *Lingress) resolveCertificate(info *tls.ClientHelloInfo) (*tls.Ce
 	return nil, nil
 }
 
-func (instance *Lingress) shutdownListener(stop support.Channel) {
+func (this *Lingress) shutdownListener(stop support.Channel) {
 	stop.Wait()
-	instance.Http.Shutdown()
-	instance.Https.Shutdown()
+	this.Http.Shutdown()
+	this.Https.Shutdown()
 }
 
-func (instance *Lingress) accessLogQueueWorker(stop support.Channel, queue chan *accessLogEntry) {
+func (this *Lingress) accessLogQueueWorker(stop support.Channel, queue chan *accessLogEntry) {
 	stopCh := support.ToChan(stop)
 	for {
 		select {
 		case entry := <-queue:
-			instance.doAccessLog(entry)
+			this.doAccessLog(entry)
 		case <-stopCh:
 			return
 		}
 	}
 }
 
-func (instance *Lingress) onAccessLog(ctx *lctx.Context) {
-	if log.IsLevelEnabled(instance.logLevelByContext(ctx)) {
+func (this *Lingress) onAccessLog(ctx *lctx.Context) {
+	if log.IsLevelEnabled(this.logLevelByContext(ctx)) {
 		entry := &accessLogEntry{
 			Context: ctx,
 		}
-		if q := instance.accessLogQueue; q != nil {
+		if q := this.accessLogQueue; q != nil {
 			q <- entry
 		} else {
-			instance.doAccessLog(entry)
+			this.doAccessLog(entry)
 		}
 	}
 }
 
-func (instance *Lingress) doAccessLog(entry *accessLogEntry) {
-	defer entry.Release()
+func (this *Lingress) doAccessLog(entry *accessLogEntry) {
+	defer func() {
+		if err := entry.Release(); err != nil {
+			this.logger.
+				WithError(err).
+				Error("Problem while releasing context.")
+		}
+	}()
 
-	entry.load(instance.InlineAccessLog)
+	entry.load(this.settings.AccessLog.Inline.Get())
 	status := 0
 	if c, ok := entry.data["client"]; ok {
 		if ps, ok := c.(map[string]interface{})["status"]; ok {
@@ -287,49 +292,24 @@ func (instance *Lingress) doAccessLog(entry *accessLogEntry) {
 			}
 		}
 	}
-	if !instance.shouldBeLogged(entry) && status < 400 {
+	if !this.shouldBeLogged(entry) && status < 400 {
 		return
 	}
 
-	f := func(sub, field string) string {
-		if v := entry.getField(sub, instance.InlineAccessLog, field); v != nil {
-			if d, ok := v.(time.Duration); ok {
-				// Because beforehand we removed that value to be microsecond. Now, for the formatter,
-				// it needs to be nanosecond again.
-				v = d * time.Microsecond
-			}
-			return fmt.Sprint(v)
-		}
-		return "-"
-	}
-
-	lvl := instance.logLevelByStatus(status)
-
-	logger := log.GetRootLogger()
-
-	event := logger.NewEvent(lvl, entry.data).
-		With("event", "accessLog").
-		Withf("message", "[%s] %s %s %s %s %q",
-			f(lctx.FieldClient, lctx.FieldClientStatus),
-			f(lctx.FieldClient, lctx.FieldClientMethod),
-			f(lctx.FieldClient, lctx.FieldClientUrl),
-			f(lctx.FieldClient, lctx.FieldClientDuration),
-			f(lctx.FieldClient, lctx.FieldClientAddress),
-			f(lctx.FieldClient, lctx.FieldClientUserAgent),
-		)
-
-	logger.Log(event, 0)
+	lvl := this.logLevelByStatus(status)
+	event := this.accessLogger.NewEvent(lvl, entry.data)
+	this.accessLogger.Log(event, 0)
 }
 
-func (instance *Lingress) shouldBeLogged(entry *accessLogEntry) bool {
-	userAgent, remotes := instance.userAgentAndRemoteOf(entry)
-	if strings.HasPrefix(userAgent, "kube-probe/") && instance.hasPrivateNetworkIp(remotes) {
+func (this *Lingress) shouldBeLogged(entry *accessLogEntry) bool {
+	userAgent, remotes := this.userAgentAndRemoteOf(entry)
+	if strings.HasPrefix(userAgent, "kube-probe/") && this.hasPrivateNetworkIp(remotes) {
 		return false
 	}
 	return true
 }
 
-func (instance *Lingress) userAgentAndRemoteOf(entry *accessLogEntry) (userAgent string, remotes []net.IP) {
+func (this *Lingress) userAgentAndRemoteOf(entry *accessLogEntry) (userAgent string, remotes []net.IP) {
 	c, ok := entry.data["client"]
 	if !ok {
 		return
@@ -355,7 +335,7 @@ func (instance *Lingress) userAgentAndRemoteOf(entry *accessLogEntry) (userAgent
 	return
 }
 
-func (instance *Lingress) hasPrivateNetworkIp(ips []net.IP) bool {
+func (this *Lingress) hasPrivateNetworkIp(ips []net.IP) bool {
 	for _, ip := range ips {
 		for _, pn := range privateNetworks {
 			if pn.Contains(ip) {
@@ -366,14 +346,14 @@ func (instance *Lingress) hasPrivateNetworkIp(ips []net.IP) bool {
 	return false
 }
 
-func (instance *Lingress) logLevelByContext(ctx *lctx.Context) level.Level {
+func (this *Lingress) logLevelByContext(ctx *lctx.Context) level.Level {
 	if err := ctx.Error; err != nil && ctx.Client.Status <= 0 {
 		ctx.Client.Status = 500
 	}
-	return instance.logLevelByStatus(ctx.Client.Status)
+	return this.logLevelByStatus(ctx.Client.Status)
 }
 
-func (instance *Lingress) logLevelByStatus(status int) level.Level {
+func (this *Lingress) logLevelByStatus(status int) level.Level {
 	if status < 500 {
 		return level.Info
 	} else if status == http.StatusBadGateway ||
@@ -385,14 +365,14 @@ func (instance *Lingress) logLevelByStatus(status int) level.Level {
 	}
 }
 
-func (instance *Lingress) onResult(ctx *lctx.Context) {
+func (this *Lingress) onResult(ctx *lctx.Context) {
 	if ctx.Result.WasResponseSendToClient() {
 		return
 	}
 
 	ctx.Stage = lctx.StagePrepareClientResponse
-	if proceed, err := instance.Proxy.Interceptors.Handle(ctx); err != nil {
-		ctx.Log().WithError(err).Error("cannot prepare error response")
+	if proceed, err := this.Proxy.Interceptors.Handle(ctx); err != nil {
+		ctx.Log().WithError(err).Error("Cannot prepare error response.")
 		return
 	} else if !proceed {
 		return
@@ -401,16 +381,16 @@ func (instance *Lingress) onResult(ctx *lctx.Context) {
 	ctx.Stage = lctx.StageSendResponseToClient
 	ctx.Client.Status = ctx.Result.Status()
 	if ctx.Result == lctx.ResultFailedWithRuleNotFound {
-		instance.Fallback.Unknown(ctx)
+		this.Fallback.Unknown(ctx)
 	} else if rr, ok := ctx.Result.(lctx.RedirectResult); ok {
-		instance.Fallback.Redirect(ctx, ctx.Client.Status, rr.Target)
+		this.Fallback.Redirect(ctx, ctx.Client.Status, rr.Target)
 	} else {
 		p := ""
 		if u, err := ctx.Client.RequestedUrl(); err == nil && u != nil {
 			p = u.Path
 		}
 		canHandleTemporary := ctx.Client.Request.Method == "GET"
-		instance.Fallback.Status(ctx, ctx.Client.Status, p, canHandleTemporary)
+		this.Fallback.Status(ctx, ctx.Client.Status, p, canHandleTemporary)
 	}
 
 	ctx.Stage = lctx.StageDone
@@ -421,20 +401,20 @@ type accessLogEntry struct {
 	data map[string]interface{}
 }
 
-func (instance *accessLogEntry) load(inline bool) {
-	instance.data = instance.AsMap(inline)
+func (this *accessLogEntry) load(inline bool) {
+	this.data = this.AsMap(inline)
 }
 
-func (instance *accessLogEntry) getField(sub string, inlined bool, field string) interface{} {
+func (this *accessLogEntry) getField(sub string, inlined bool, field string) interface{} {
 	if sub == "" {
-		return instance.data[field]
+		return this.data[field]
 	}
 
 	if inlined {
-		return instance.data[sub+"."+field]
+		return this.data[sub+"."+field]
 	}
 
-	if sm, ok := instance.data[sub].(map[string]interface{}); ok {
+	if sm, ok := this.data[sub].(map[string]interface{}); ok {
 		return sm[field]
 	}
 
