@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/echocat/lingress/context"
 	"github.com/echocat/lingress/rules"
-	"github.com/echocat/lingress/support"
 	"github.com/echocat/lingress/value"
 	"net"
 	"net/http"
@@ -17,78 +16,36 @@ func init() {
 }
 
 var (
-	defaultMaxAge = value.Duration(24 * time.Hour)
+	defaultMaxAge = value.NewDuration(24 * time.Hour)
 )
 
-type CorsInterceptor struct {
-	AllowedOriginsHost rules.ForcibleHostPatterns
-	AllowedMethods     rules.ForcibleMethods
-	AllowedHeaders     rules.ForcibleHeaderNames
-	AllowedCredentials value.ForcibleBool
-	MaxAge             value.ForcibleDuration
-	Enabled            value.ForcibleBool
-}
+type CorsInterceptor struct{}
 
 func NewCorsInterceptor() *CorsInterceptor {
-	return &CorsInterceptor{
-		AllowedOriginsHost: rules.NewForcibleHostPatterns(rules.HostPatterns{}, false),
-		AllowedMethods:     rules.NewForcibleMethods(rules.Methods{}, false),
-		AllowedHeaders:     rules.NewForcibleHeaders(rules.HeaderNames{}, false),
-		AllowedCredentials: value.NewForcibleBool(value.True, false),
-		MaxAge:             value.NewForcibleDuration(defaultMaxAge, false),
-		Enabled:            value.NewForcibleBool(value.False, false),
-	}
+	return &CorsInterceptor{}
 }
 
-func (instance *CorsInterceptor) Name() string {
+func (this *CorsInterceptor) Name() string {
 	return "cors"
 }
 
-func (instance *CorsInterceptor) HandlesStages() []context.Stage {
+func (this *CorsInterceptor) HandlesStages() []context.Stage {
 	return []context.Stage{context.StageEvaluateClientRequest, context.StagePrepareClientResponse}
 }
 
-func (instance *CorsInterceptor) Handle(ctx *context.Context) (proceed bool, err error) {
+func (this *CorsInterceptor) Handle(ctx *context.Context) (proceed bool, err error) {
 	switch ctx.Stage {
 	case context.StageEvaluateClientRequest:
-		return instance.enableCors(ctx)
+		return this.enableCors(ctx)
 	case context.StagePrepareClientResponse:
-		return instance.forceCorsHeaders(ctx)
+		return this.forceCorsHeaders(ctx)
 	}
 	return true, nil
 }
 
-func (instance *CorsInterceptor) RegisterFlag(fe support.FlagEnabled, appPrefix string) error {
-	fe.Flag("cors.enabled", "If set if will be used if annotation lingress.echocat.org/cors is absent. If this value is prefix with ! it overrides everything regardless what was set in the annotation.").
-		PlaceHolder(fmt.Sprint(instance.Enabled)).
-		Envar(support.FlagEnvName(appPrefix, "CORS_ENABLED")).
-		SetValue(&instance.Enabled)
-	fe.Flag("cors.allowedOriginHosts", "Host patterns which are allowed to be referenced by Origin headers. '*' means all.").
-		PlaceHolder(instance.AllowedOriginsHost.String()).
-		Envar(support.FlagEnvName(appPrefix, "CORS_ALLOWED_ORIGIN_HOSTS")).
-		SetValue(&instance.AllowedOriginsHost)
-	fe.Flag("cors.allowedMethods", "Methods which are allowed to be referenced.").
-		PlaceHolder(fmt.Sprint(instance.AllowedMethods)).
-		Envar(support.FlagEnvName(appPrefix, "CORS_ALLOWED_METHODS")).
-		SetValue(&instance.AllowedMethods)
-	fe.Flag("cors.allowedHeaders", "Headers which are allowed to be referenced. '*' means all.").
-		PlaceHolder(fmt.Sprint(instance.AllowedHeaders)).
-		Envar(support.FlagEnvName(appPrefix, "CORS_ALLOWED_HEADERS")).
-		SetValue(&instance.AllowedHeaders)
-	fe.Flag("cors.allowedCredentials", "Credentials are allowed.").
-		PlaceHolder(fmt.Sprint(instance.AllowedCredentials)).
-		Envar(support.FlagEnvName(appPrefix, "CORS_ALLOWED_CREDENTIALS")).
-		SetValue(&instance.AllowedCredentials)
-	fe.Flag("cors.maxAge", "How long the response to the preflight request can be cached for without sending another preflight request").
-		PlaceHolder(fmt.Sprint(instance.MaxAge)).
-		Envar(support.FlagEnvName(appPrefix, "CORS_MAX_AGE")).
-		SetValue(&instance.MaxAge)
-	return nil
-}
-
-func (instance *CorsInterceptor) enableCors(ctx *context.Context) (proceed bool, err error) {
+func (this *CorsInterceptor) enableCors(ctx *context.Context) (proceed bool, err error) {
 	optionsCors := rules.OptionsCorsOf(ctx.Rule)
-	if !instance.Enabled.Select(optionsCors.Enabled).GetOr(false) {
+	if !ctx.Settings.Cors.Enabled.Select(optionsCors.Enabled).GetOr(false) {
 		return true, nil
 	}
 	origin, err := ctx.Client.Origin()
@@ -102,7 +59,7 @@ func (instance *CorsInterceptor) enableCors(ctx *context.Context) (proceed bool,
 	if sHost, _, err := net.SplitHostPort(host); err == nil {
 		host = sHost
 	}
-	if !instance.AllowedOriginsHost.Evaluate(optionsCors.AllowedOriginsHost, nil).Matches(host) {
+	if !ctx.Settings.Cors.AllowedOriginsHost.Evaluate(optionsCors.AllowedOriginsHost).Matches(host) {
 		ctx.Client.Response.Header().Set("X-Cors-Hint", "origin-ignored")
 		return true, nil
 	}
@@ -116,7 +73,7 @@ func (instance *CorsInterceptor) enableCors(ctx *context.Context) (proceed bool,
 	return true, nil
 }
 
-func (instance *CorsInterceptor) deleteHeaders(h http.Header) {
+func (this *CorsInterceptor) deleteHeaders(h http.Header) {
 	h.Del("Access-Control-Allow-Origin")
 	h.Del("Access-Control-Allow-Credentials")
 	h.Del("Access-Control-Allow-Methods")
@@ -124,10 +81,10 @@ func (instance *CorsInterceptor) deleteHeaders(h http.Header) {
 	h.Del("Access-Control-Max-Age")
 }
 
-func (instance *CorsInterceptor) forceCorsHeaders(ctx *context.Context) (proceed bool, err error) {
+func (this *CorsInterceptor) forceCorsHeaders(ctx *context.Context) (proceed bool, err error) {
 	proceed = true
 	cors := rules.OptionsCorsOf(ctx.Rule)
-	enabled := instance.Enabled.Select(cors.Enabled)
+	enabled := ctx.Settings.Cors.Enabled.Select(cors.Enabled)
 	h := ctx.Client.Response.Header()
 
 	if enabled.GetOr(false) {
@@ -136,24 +93,24 @@ func (instance *CorsInterceptor) forceCorsHeaders(ctx *context.Context) (proceed
 			return false, fmt.Errorf("cannot enable cors: %w", oErr)
 		}
 		if origin == nil {
-			instance.deleteHeaders(h)
+			this.deleteHeaders(h)
 			return
 		}
 		host := origin.Host
 		if sHost, _, err := net.SplitHostPort(host); err == nil {
 			host = sHost
 		}
-		if !instance.AllowedOriginsHost.Evaluate(cors.AllowedOriginsHost, nil).Matches(host) {
-			instance.deleteHeaders(h)
+		if !ctx.Settings.Cors.AllowedOriginsHost.Evaluate(cors.AllowedOriginsHost).Matches(host) {
+			this.deleteHeaders(h)
 			return
 		}
 		h.Set("Access-Control-Allow-Origin", origin.String())
-		h.Set("Access-Control-Allow-Credentials", fmt.Sprint(instance.AllowedCredentials.Evaluate(cors.AllowedCredentials, true)))
-		h.Set("Access-Control-Allow-Methods", instance.AllowedMethods.Evaluate(cors.AllowedMethods, nil).String())
-		h.Set("Access-Control-Allow-Headers", instance.AllowedHeaders.Evaluate(cors.AllowedHeaders, nil).String())
-		h.Set("Access-Control-Max-Age", strconv.Itoa(instance.MaxAge.Evaluate(cors.MaxAge, defaultMaxAge).AsSeconds()))
+		h.Set("Access-Control-Allow-Credentials", fmt.Sprint(ctx.Settings.Cors.AllowedCredentials.Evaluate(cors.AllowedCredentials).GetOr(true)))
+		h.Set("Access-Control-Allow-Methods", ctx.Settings.Cors.AllowedMethods.Evaluate(cors.AllowedMethods).String())
+		h.Set("Access-Control-Allow-Headers", ctx.Settings.Cors.AllowedHeaders.Evaluate(cors.AllowedHeaders).String())
+		h.Set("Access-Control-Max-Age", strconv.Itoa(ctx.Settings.Cors.MaxAge.EvaluateOr(cors.MaxAge, defaultMaxAge).AsSeconds()))
 	} else if enabled.IsForced() {
-		instance.deleteHeaders(h)
+		this.deleteHeaders(h)
 	}
 
 	return
