@@ -3,7 +3,7 @@ package rules
 import "github.com/echocat/lingress/rules/tree"
 
 type ByPath struct {
-	values *tree.Tree
+	values *tree.Tree[Rule]
 
 	onAdded   OnAdded
 	onRemoved OnRemoved
@@ -11,58 +11,48 @@ type ByPath struct {
 
 func NewByPath(onAdded OnAdded, onRemoved OnRemoved) *ByPath {
 	result := &ByPath{
-		values:    tree.New(),
+		values:    tree.New[Rule](),
 		onAdded:   onAdded,
 		onRemoved: onRemoved,
 	}
 
-	result.values.OnAdded = func(path []string, element interface{}) {
-		if r, ok := element.(Rule); ok {
-			onAdded(path, r)
-		}
+	result.values.OnAdded = func(path []string, r Rule) {
+		onAdded(path, r)
 	}
-	result.values.OnRemoved = func(path []string, element interface{}) {
-		if r, ok := element.(Rule); ok {
-			onRemoved(path, r)
-		}
+	result.values.OnRemoved = func(path []string, r Rule) {
+		onRemoved(path, r)
 	}
 	return result
 }
 
-func (instance *ByPath) All(consumer func(Rule) error) error {
-	return instance.values.All(func(value interface{}) error {
-		return consumer(value.(Rule))
-	})
+func (this *ByPath) All(consumer func(Rule) error) error {
+	return this.values.All(consumer)
 }
 
-func (instance *ByPath) Find(path []string) (Rules, error) {
-	if plain, err := instance.values.Find(path); err != nil {
+func (this *ByPath) Find(path []string) (Rules, error) {
+	if plain, err := this.values.Find(path); err != nil {
 		return nil, err
 	} else {
 		return rules(plain), err
 	}
 }
 
-func (instance *ByPath) Put(r Rule) error {
-	return instance.values.Put(r.Path(), r)
+func (this *ByPath) Put(r Rule) error {
+	return this.values.Put(r.Path(), r)
 }
 
-func (instance *ByPath) Remove(predicate Predicate) error {
-	return instance.values.Remove(func(path []string, element interface{}) bool {
-		if r, ok := element.(Rule); ok {
-			return predicate(path, r)
-		} else {
-			return false
-		}
+func (this *ByPath) Remove(predicate Predicate) error {
+	return this.values.Remove(func(path []string, r Rule) bool {
+		return predicate(path, r)
 	})
 }
 
-func (instance *ByPath) HasContent() bool {
-	return instance.values.HasContent()
+func (this *ByPath) HasContent() bool {
+	return this.values.HasContent()
 }
 
-func (instance *ByPath) Clone() *ByPath {
-	result := NewByPath(instance.onAdded, instance.onRemoved)
-	result.values = instance.values.Clone()
+func (this *ByPath) Clone() *ByPath {
+	result := NewByPath(this.onAdded, this.onRemoved)
+	result.values = this.values.Clone()
 	return result
 }
