@@ -1,36 +1,52 @@
 package rules
 
-import "fmt"
+import (
+	"fmt"
+	"slices"
+)
 
 type Rules interface {
 	Get(int) Rule
 	Len() int
 	Any() Rule
+	AnyFilteredBy(path []string) Rule
 }
 
-type rules []interface{}
+type rules []Rule
 
-func (instance rules) Get(i int) Rule {
-	return instance[i].(Rule)
+func (this rules) Get(i int) Rule {
+	return this[i]
 }
 
-func (instance rules) Len() int {
-	if instance == nil {
+func (this rules) Len() int {
+	if this == nil {
 		return 0
 	}
-	return len(instance)
+	return len(this)
 }
 
-func (instance rules) Any() Rule {
-	if len(instance) > 0 {
-		return (instance)[0].(Rule)
+func (this rules) Any() Rule {
+	if len(this) > 0 {
+		return (this)[0]
 	}
 	return nil
 }
 
-func (instance rules) String() string {
+func (this rules) AnyFilteredBy(path []string) Rule {
+	for _, candidate := range this {
+		if candidate.PathType() != PathTypeExact {
+			return candidate
+		}
+		if slices.Equal(path, candidate.Path()) {
+			return candidate
+		}
+	}
+	return nil
+}
+
+func (this rules) String() string {
 	result := ""
-	for i, r := range instance {
+	for i, r := range this {
 		if i > 0 {
 			result += "\n"
 		}
