@@ -11,26 +11,26 @@ type Statistics struct {
 	totalDuration  time.Duration
 }
 
-func (instance *Statistics) MarkUsed(duration time.Duration) {
-	atomic.AddUint64(&instance.numberOfUsages, 1)
-	atomic.AddInt64((*int64)(&instance.totalDuration), int64(duration))
+func (this *Statistics) MarkUsed(duration time.Duration) {
+	atomic.AddUint64(&this.numberOfUsages, 1)
+	atomic.AddInt64((*int64)(&this.totalDuration), int64(duration))
 }
 
-func (instance *Statistics) NumberOfUsages() uint64 {
-	return atomic.LoadUint64(&instance.numberOfUsages)
+func (this *Statistics) NumberOfUsages() uint64 {
+	return atomic.LoadUint64(&this.numberOfUsages)
 }
 
-func (instance *Statistics) TotalDuration() time.Duration {
-	return time.Duration(atomic.LoadInt64((*int64)(&instance.totalDuration)))
+func (this *Statistics) TotalDuration() time.Duration {
+	return time.Duration(atomic.LoadInt64((*int64)(&this.totalDuration)))
 }
 
-func (instance *Statistics) MarshalJSON() ([]byte, error) {
+func (this *Statistics) MarshalJSON() ([]byte, error) {
 	v := struct {
 		NumberOfUsages uint64        `json:"numberOfUsages"`
 		TotalDuration  time.Duration `json:"totalDuration"`
 	}{
-		NumberOfUsages: instance.NumberOfUsages(),
-		TotalDuration:  instance.TotalDuration() / time.Microsecond,
+		NumberOfUsages: this.NumberOfUsages(),
+		TotalDuration:  this.TotalDuration() / time.Microsecond,
 	}
 	return json.Marshal(v)
 }
